+++ core/src/main/java/org/apache/calcite/sql2rel/RelFieldTrimmer.java
@@ -481,15 +481,8 @@ public class RelFieldTrimmer implements ReflectiveVisitor {
     final int fieldCount = rowType.getFieldCount();
     final RelNode input = project.getInput();
 
-    boolean containsSubQuery = false;
-    for (RexNode node : project.getProjects()) {
-      if (RexUtil.containsSubQuery(node)) {
-        containsSubQuery = true;
-        break;
-      }
-    }
     // Do not trim Project's fields before SubQueryRemoveRule applies.
-    if (containsSubQuery) {
+    if (RexUtil.containsSubQuery(project.getProjects())) {
       return result(project, Mappings.createIdentity(fieldCount));
     }