+++ core/src/main/java/org/apache/calcite/tools/RelBuilder.java
@@ -1908,19 +1908,9 @@ public class RelBuilder {
       fieldNameList.add(null);
     }
 
-    // Do not merge projections when top project contains RexSubQuery
-    boolean containsSubQuery = false;
-    for (RexNode node : nodes) {
-      if (RexUtil.containsSubQuery(node)) {
-        containsSubQuery = true;
-        break;
-      }
-    }
-
     bloat:
     if (frame.rel instanceof Project
-        && config.bloat() >= 0
-        && !containsSubQuery) {
+        && config.bloat() >= 0) {
       final Project project = (Project) frame.rel;
       // Populate field names. If the upper expression is an input ref and does
       // not have a recommended name, use the name of the underlying field.
@@ -1940,6 +1930,10 @@ public class RelBuilder {
       if (newNodes == null) {
         // The merged expression is more complex than the input expressions.
         // Do not merge.
+        break bloat;
+      }
+      // Do not merge projections when top project contains RexSubQuery
+      if (RexUtil.containsSubQuery(nodes)) {
         break bloat;
       }